Fix incorrect focusedPlaceholderColor
Given that a user enter and removes text in TextField the composable for
placeholder decoration will be recreated, which contains the state of the
placeholderColor. When recreating the placeholderColor the focused property
will be to reset to default, false, this causes TextField to use the color of
an unfocused placeholder when infact the TextField is focused.

Fixes: b/313669163