Add RadioButton component for selection controls like 'Radio'
Radio buttons differ from toggle buttons in that they are 'selectable' rather than 'toggleable'. This is especially noticeable under TalkBack. Separated the RadioButton/ToggleButton code and tests. Also, moved the 'Radio' control into SelectionControls and the Checkbox/Switch into ToggleControls.

Bug: 313432610
Test: Added RadioButton tests and updated ToggleButton/ToggleControl tests.
Relnote: "We have added RadioButton/SplitRadioButton as containers for selection controls, such as the Radio control. This differs from the existing ToggleButton in that RadioButton is selectable (and operates within a selection group) whereas ToggleButton is toggleable (and is independent)."
Change-Id: I612758b8625c45756e6a56d31c1008daac45fba3