Relax cubic precision a bit
In a recent change, we started throwing an exception instead
of returning the input fraction when our computations produce
a NaN. This caused a regression on Wear when attempting to
transform an input value very close to 0.0f. Our fix here is
to slightly relax our epsilon to treat the output of
computations close to 0 or 1 as 0 or 1.

We also add a test to validate values close to 0, chosen
to be both above and below the epsilon.

Fixes: 314313566
Test: EasingTest
Change-Id: If50a63be9aa58a21011e9b23ccf7f951e35b3e24