[GH] Fix incorrect focusedPlaceholderColor
Given that a user enter and removes text in TextField the composable for placeholder decoration will be recreated, which contains the state of the placeholderColor. When recreating the placeholderColor the focused property will be to reset to default, false, this causes TextField to use the color of an unfocused placeholder when infact the TextField is focused.

## Issues Fixed

Fixes: [b/313669163](https://issuetracker.google.com/issues/313669163)

Tests: none, I'm unable to run the tests suite. Update in manual would be appreciated.

This is an imported pull request from https://github.com/androidx/androidx/pull/631.

Resolves #631
Github-Pr-Head-Sha: 033fd3be8921b23af93975e0a5f74256c48aa360
GitOrigin-RevId: 2bbfee4ced5115d6f7e5621c7e7e459b37c44f9e
Change-Id: If0e1fc87d798f817985083600b1122a294799e22