Add a new COORDINATE_SYSTEM_SENSOR option
This option can be used along with the new OverlayEffect API. To draw an overlay Canvas correctly, we need to first transform the Canvas to camera sensor coordinate system. In that case, we also need the MlKitAnalyzer to produce results in camera sensor coordinate system. In this CL:

- Add a new ImageAnalysis.COORDINATE_SYSTEM_SENSOR option.
- In MlKitAnalyzer, ignore the additional transformation if this option is used. Only revert the sensor to buffer transformation.

Relnote: add a ImageAnalysis#COORDINATE_SYSTEM_SENSOR. When this is used, the MlKitAnalyzer return coordinates in the camera sensor coordinates system.

Bug: 310260193
Test: manual test and ./gradlew bOS
Change-Id: I3ec61b28688aafb8398b7ddf7aa5f9325fc73a2d