Fix Spacer bug with mutation
For Spacer with no Modifiers we need to set minimum dimension,
however was recently changed to be done after checking LayoutParams
which can be null in valid case (in mutation, when parent is not
there/doesn't have a diff. For those cases, we should always first
apply minimum dimension (preserving the old behaviour) and then
update LayoutParams if needed.

Additionally, resetting LayoutParams to 0 for LinearDimension before
adding View to its parent so that's correctly propagated and not
have previous value in mutation.

Bug: N/A
Test: Added
Change-Id: Iff36850372a29e11f9049ef3386a3e11a07d7f82