Skip AutoboxingStateCreation check for Java code
This inspection wasn't designed with Java in mind and a user has
reported that it throws an exception when linting their Java code,
where they call mutableStateOf.

We don't have enough nullity information for the generic type of a
user's state when they call mutableStateOf in Java to guarantee that
they could make the change proposed by this inspection, so let's avoid
any crashes by intentionally skipping the check for all Java files.

Fixes: b/314093514
Test: testTrivialMutableStateOf_notReportedInJava
Change-Id: I02d101a739a2d2f953f7c277e74ac078188272c1