Add tests for using inline types in Java
When using Kotlin inline types directly in Kotlin kotlinc replaces them
with the underlying type and when used in Java javac doesn't replace
them. But the KSP side of XProcessing can't tell the two and always
replaces them with the underlying type when generating type names
for inline types.

This CL just documents current behavior with tests. Solving this may
be difficult as an XType doesn't know if it's being used in Java
or Kotlin.

Test: XTypeTest.kt
Change-Id: Iec4dbc6054ed74cff3f62f0bb3b827644f47758c