Make `currentCompoundHashKey` unique in more cases
Prior to this change calling a composable function multiple times
in the same basic block would received the same compound hash key
as the key was a hash of the groups created to track composition
from the root to the call. Calling the same function would produce
a group with exactly the same key which would then create a
duplicate compound hash key.

With this change the number of non-movable groups (or replaceable
groups) prior to the new replaceable group is factored in as part
of the hash key to ensure that hash keys are much more likely to
be unique in these cases.

Moveable groups (generated by calls to `key`) are unaffected by
this change. The uniqueness of the compound keys depends on the
uniquness of the value(s) passed to `key`.

Fixes: 287537290
Test: ./gradlew :compose:r:r:tDUT

Change-Id: I65507c0d4f5ac5d92db2619b284b1f3266c04e2f