Reduce the size of libandroidx.graphics.path.so by 96%
The statically linked libstdc++ was bringing over a ton of code that
is not used by this library, even after removing all non-header-only
usages of STL. The last remaining usage was the allocating new,
which this CL removes. We now manually malloc/free the storage, and
call the constructor using the placement new operator, while
invoking the destructor manually.

With these changes, we can remove stdlib++ at link time, which
drops the .so from 351 KiB down to 14 KiB for aarch64.

RelNote: Reduced the size of libandroidx.graphics.path.so by 96%
Test: PathIteratorTest, PathEasingTest
Change-Id: I7139702def79ecfcf3d4dfe7d36f833230cf754b