Properly handle ClassCastException in SavedStateHandle when calling get() with the incorrect type
When using SavedStateHandle, if you attempt to retrieve a value with the
incorrect class type, it is casted to the correct class type and ends up
throwing a ClassCastException. But inside `catch` block, we call `remove<T>(key)`,
which will throw another ClassCastException, because `remove` method also tries to cast the existing value to the given type.
This commit changes `T` to `Any?` to avoid ClassCastException.