Commonize Room auto migration validation and initialization.
This change moves auto migration spec validation and auto migration initialization into common code. Generated auto migration related functions had to be mirrored into common code using kotlin.reflect.KClass instead java.lang.Class with a backwards compatible path for old generated code. Reflection assignment checks had to be migrated to an actual/expect function that only does FQN comparison in native which is not quite equivalent as that would fail if a sub-class of the spec is provided, but this might be an unrealistic case. Why would someone provide a subclass of their @ProvidedAutoMigrationSpec? If this is an issue we can enforce @ProvidedAutoMigrationSpec to be final in processor (b/317210564). Regardless of, there is no reflection in native code, so this will do.

Other related changes:
 * Added onPostMigrate(connection: SQLiteConnection) but note that such callback is not invoked yet (need to change gen code).
 * Migrated relevant concrete collection types in gen code to use Kotlin's Stdlib builders as those are KMP compatible.
 * Add XCodeBlock.ofKotlinClassLiteral() to emit KClass literals, note that in Java it converts the java.lang.Class using JvmClassMappingKt but such case should never be generated.

Bug: 299168035
Bug: 309993778
Test: Existing
Relnote: "Add onPostMigrate(connection: SQLiteConnection) to AutoMigrationSpec"
Change-Id: I3b7c9219f6ba79056633530b052d3ce54f013940