Updates material3 components to expose a nullable MutableInteractionSource
Components that previously exposed an API such as:

interactionSource: MutableInteractionSource = remember { MutableInteractionSource() }

Have been updated to instead expose a nullable MutableInteractionSource that defaults to null. This allows for some components to never create and remember a MutableInteractionSource, and allows for some other components to lazily create one only when needed. Some components always need a MutableInteractionSource currently, so if null they will just create a new one internally - so for those cases there are no changes, but it allows for future optimizations.

Test: updateApi, existing tests
Bug: b/298048146
Relnote: "Material3 components exposing a MutableInteractionSource in their API have been updated to now expose a nullable MutableInteractionSource that defaults to null. There are no semantic changes here: passing null means that you do not wish to hoist the MutableInteractionSource, and it will be created inside the component if needed. Changing to null allows for some components to never allocate a MutableInteractionSource, and allows for other components to only lazily create an instance when they need to, which improves performance across these components. If you are not using the MutableInteractionSource you pass to these components, it is recommended that you pass null instead. It is also recommended that you make similar changes in your own components."
Change-Id: I41abb601499b4a735b6302b96cdc1f0d066dbbdc