Remove root view focus check when executing keyboard commands
Background; A while ago a focus check was added to `processInputCommands` to skip executing commands at the end of the frame if the owner view has already lost its focus. This was addressing a problem where a focus switch from a TextField to EditText would cause Keyboard to flicker.

Problem; A new change to the Compose Focus system now correctly releases the focus from Compose owner view when there are no longer focused Composables in its control. However, this invalidates `TextInputServiceAndroid#processInputCommands`'s assumption that Compose owner view would always have focus unless it's transferred to another Android view.

Disposing a TextField, or clearing focus from a TextField, no longer hides the keyboard which is a serious problem. This CL fixes this problem by removing the previously added focus check.

Note; THIS BRINGS BACK THE INTEROP BUG. This CL unfortunately would bring back the keyboard problem when interacting between TextField and EditText. However due to the much higher priority of the original bug that's addressed in this CL, we will leave solving this re-introduced bug as a future work.

Bug: 318424446
Bug: 313375840
Test: :compose:foundation:foundation:cAT
Change-Id: Ibc5dc78927d75c8220a62587e9ce8325929585ba