Merge room-ktx into room-runtime
And very importantly add Coroutines as a dependency on room-runtime.

Also add an explicit dep to room-ktx to certain projects since atomic group constraints are not enforced via project() dependencies.

Bug: 317120607
Relnote: "RoomDatabase Kotlin extensions, withTransaction(), invalidationTrackerFlow() and others have been moved from `room-ktx` to `room-runtime`. The artifact `room-ktx` is now empty and is not needed for Coroutines support in Room, and specifically not needed when defining suspend functions in DAOs."
Test: ./gradlew checkApi
Change-Id: Ib7619848956632f51608db6d3e0f40f6782c6251