Optimize fp32<>fp16 conversions
This change is useful to optimize the creation and manipulation
of non-sRGB colors. In non-sRGB colors, the color channels are
encoded in fp16 inside a 64-bit word.

The optimization triggers only on API 30+ on Android, where
conversions to/from fp16 are implemented as an ART intrinsic,
which relies in turn on aarch64 instructions.

On a Pixel 6 runnin API 34, the improvements are as follows:

- Lerping wide colors:
  237,363ns -> 212,333ns, a 10.5% improvement

- Creating/extracting wide colors via Color:
  284,740ms -> 161,162ns, a 43.4% improvement

Test: ColorTest
RelNote: Improved performance of encoding/decoding wide gamut colors
Change-Id: Ic37359670a50a196908d3f2aa0415c73154066b2