Migrate #setLaunchingActivityStack to bundle
This CL migrate all #setLunchingActivityStack APIs to bundle
to compat with ActivityOptionsCompat usages.

This CL also removes #setLaunchingActivity(Activity) seems
the behavior is confusing. Users may not know if an
ActivityStack is attached with this API. Users should migrate
to use ActivityEmbeddingController#getActivityStack(Activity)
to get an ActivityStack, and pass the ActivityStack to
 #setLaunchingActivityStack.

Test: ./gradlew window:window:test
Bug: 295993745
Relnote: "
  - Migrate #setLaunchingActivityStack to Bundle to compat with
    ActivityOptionsCompat usages.
  - Users should pass activityOptions.toBundle instead of
    ActvityOptions itself.
  - Remove #setLaunchingActivityStack(Activity). Users should
    migrate to use
    ActivityEmbeddingController#getActivityStack(Activity)
    to get an ActivityStack, and pass the ActivityStack to
    #setLaunchingActivityStack.
"

Change-Id: Ie0ccc35e5459c9031d6890185d6472388eb68320