Add check for layout with large dimensions.
Bug: 316365522

When a layout uses extremely large values in its calculated
size, such as Constraints.Infinity, it can cause problems
in other layouts. It is very difficult to track down the
source of the problem layout, as the problem is usually
detected in a parent when it tries to use the large
dimension in Constraints.

This CL adds a check whenever a child uses layout()
to ensure that the dimension isn't absurdly large.
This should eliminate any common problem with using
Constraints.Infinity, such as directly using
Constraints.maxWidth or a calculation from a maximum
dimension.

Test: ran existing layout tests. Created new tests.
Change-Id: I339a906701f382f7e043be65de57af6b9838e695