Add widgetBackground color role to Glance Theme.
Adds a new widget specific color role to Glance Theme called widgetBackground.
This is different from regular background in that it tends to have more
color in, required when being shown on home screens as opposed to as the
background of an app.


Bug: 321277130
Test: updated theme test widget
Relnote: Adds a new widgetBackground color role to Glance themes.

Change-Id: Ia2ab8f629fbed26c42074ab52e39bc81b3cf8792