Protect trackingEnabled from being set/unset multiple times
Setting the trackingEnabled flag has the side effect of adding
OnFrameListeners. Setting it multiple times adds multiple listeners
(which is not obvious for a simple boolean flag). This change simply
checks the current value and noops if the new value is the same.

Note that this is only a problem on API23 and below; there is special
logic related to FrameMetrics that avoids the problem in a different
way.

Also, minor cleanup removing implementation of getRootView() since
an equivalent and simpler version of it already exists on View.

Bug: 244619257
Bug: 244465700
Test: Added new test. JankStats passes on APIs below and above 24.
Change-Id: Idcff659680df1ac763b2383c29279c3e5cd763d1