Support select till the end of the file / till the start of the file keyboard actions on Windows (#989)
This PR is addressing following task

https://youtrack.jetbrains.com/issue/COMPOSE-784/Enable-selecting-till-the-end-of-the-file-till-the-beginning-of-the-file-in-Windows

(As of now, we do not support the standard way of selecting from current
position till the end of the text / till the beginning of the text (that
is - Shift / Ctrl / End and Shift / Ctrl / Home respectively) - but in
MacOS we actually do. This goal of this task is to introduce such
support)

Main changes are following:
* In KeyMapping.skikoMain.kt additional mapping is introduced and used
instead `defaultKeyMapping` on. desktop (which can be done later on,
it's just that we usually - this is at least my understanding - avoid
intrusive changes to commonMain as the very first step if possible)
* SelectionTest rewritten so there'll be less clutter whenever we are
adding new one