Multiplatform support for lifecycle-common
This change introduces multiplatform support for the lifecycle-common module by exposing the Lifecycle API to common, ensuring compatibility with published versions on Android. It will allow bring multiplatform support for dependent modules and use it as multiplatform abstraction for lifecycle events.

Please note that `MainThread` annotation in common requires a new version of `androidx.annotation` package.

Test: Run existing tests
Change-Id: Iedc79ae83a3a9f2ed6b6de16f002a4c91abdf4f9

# Conflicts:
#	settings.gradle