API Addition - entryGroupId + Cleanup
This adds the ability for the entries to be deduplicated, and
furthermore, for the entries to be grouped, for display purposes.
This is done via setting every entry to have specified entryGroupIds,
often username or title, or in the case of the custom type, title by
default, otherwise whatever is passed in.

Some fixes were added to these API changes in general - namely
improvements on the deprecation.

Bug: 319454328
Test: Followed up with JUnit Tests
Relnote: "Adds a entryGroupId bit to the credential entries shown in
our selector"

Change-Id: Id995c01e337ecbef082fc2cbc1f576852982bb8d