Add Expect APIs to Kruth
Note to API council - this is an unpublished library, so please ignore
this for API review. API tracking is only enabled to check against
compatibility with the public API of Truth.

Expect adds a way to do soft asserts in Kruth. These APIs are kept
JVM-only for now, as they have an unfortunate dependency on JUnit
TestRule, but it should be possible to commonize the logic and keep
binary compat by splitting the implementation in a future change.

This change also more closely aligns with the way Truth is
implemented. These are JVM-only because they depend on JUnit TestRule,
but presumably we could figure out a way to implement these on top of
kotlin.test or other KMP-compatible test engines to bring the same
functionality. Since it is strictly additive, I wanted to start by
matching the API in just JVM.

In order to get the original Truth tests passing it was required to fix
b/296945032 to align the way stacktraces are cleaned up / passed around
in the resulting exceptions. It was also necessary to fix the return
type of FailureStrategy#fail so that it does not return Nothing (forced
to throw), as soft assert depends on the ability to defer this.

NOTE: Truth has logic to clean stack traces via StackTraceCleaner, but
there are only JVM-apis to modify stack traces of thrown exceptions,
since they work differently on different platforms. Also, the primary
way of constructing AssertionErrorWithFacts, which instantiates it with
an empty stack trace has been deferred to a future CL to keep this one
smaller for review.

Fixes: 296945032
Relnote: "N/A"
Test: ANDROIDX_PROJECTS=INFRAROGUE ./gradlew kruth:kruth:allTests -Pandroidx.enabled.kmp.target.platforms="+LINUX,+MAC" --rerun-tasks
Change-Id: I084ce7c5373dd0e0c9c58f10c4c85599f65ea015