Modify the offset calculation for adding item after the last item.
When the focus is on the last item and an item is added before the focus item. If the previous item of the same row cannot be found, the offset of the adjacent previous item should be used directly. If found, the offset of the item should be added with the size of the item and space as the current offset.

Test: run locally, no crash