Add new LoadStates and CombinedLoadStates API
Added new isIdle and hasError APIS which returns whether the LoadStates or CombinedLoadStates is in NotLoading(idle) states or is in Error(hasError) states.

Test: ANDROIDX_PROJECTS=INFRAROGUE ./gradlew paging:paging-common:allTest --no-configuration-cache
Test: ./gradlew paging:paging-testing:test
Bug: n/a
Relnote: "Added new LoadStates and CombinedLoadStates helper methods to check if LoadStates is in Error or NotLoading state. Also added a new API that awaits on a LoadStateFlow until a load has settled into either NotLoading or Error state"
Change-Id: Id6c67a6ba3805130d1a08840bad38e179604f786