Ensures BTF2 cursor is scrolled back into view when typing.
If the cursor is scrolled out of view, either from the field's internal
scrollable or something outside it, and the cursor is moved or text is
entered, the cursor should be scrolled back into view as long as the
field is focused. This change makes that happen for BTF2.

This supersedes my original impl, aosp/2178021.

Bug: b/237190748
Test: TextFieldScrollTest
Relnote: "`BasicTextField2` now keeps the cursor in view while typing
    when it has been scrolled out of view or would move out of view due
    to input."
Change-Id: Ieb85691dd1a7cf98ab5fc188721d4e4475aec762