Do not allow multiple started OnBackPressCallbacks
When you have a started OnBackPressedCallback, it is possible for
another system back to be dispatched and send a signal to another
onBackPressedCallback. Having 2 callbacks running at the same time can
caused undefined behavior and should be prevents.

This change automatically cancels the previous callback when another one
is started.

RelNote: "If another back event is started during a predictive back
event, the predictive back event is automatically cancelled and the new
back event is the only one running."
Test: added test

Change-Id: I3482edb367c56c3feba833c7ba8cedf3293467bb