Add lint rule for calling onBackPressed in callback
You should not call onBackPressed on the Activity or
OnBackPressedDisptacher inside of the handledOnBackPressed of an
OnBackPressedCallback. It is an anti-pattern and should be prevented.

This change adds a lint rule to ensure to call it out at compile time.

RelNote: "There is a now a lint rule to detect calls to onBackPressed
inside of OnBackPressedCallback.handledOnBackPressed."
Test: added test
Bug: 287505200

Change-Id: I3b5d8da34b9d8b43d765a05b8cc387aafc186df0