Support sizeTransform for Navigation Compose
Adding support for passing a SizeTransform to a composable or navigation
function when building the NavGraph. This will allow developers to
specify the SizeTransform for a specfic destination.

Note the sizeTransform is determined by the target destination, which is
normally the destination that runs the enter and popEnter transitions
and the default value is `null`.

RelNote: "You can now specify the `SizeTranform` for your transitions in
Navigation Compose by defining them as part of the initialization for
the `composable` and/or `navigation` functions."
Test: work in progress
Bug: 296912651

Change-Id: I91062221a47601b40ced80dc1a31f170351e9f5a