Remove mIsNotDisplayedBySystem from visibilityConfig.
This logic is reversed with others. (the default is visible and the
default of other settings is invisible). So this will have a totally
reverse logic with others when supporting "AND" condition.

Previously, if the user didn't call SetNotDisplayedBySystem, and we got
the default ‘true’ value,  it only means the user doesn’t care whether
the schema is visible to the system. Not the user only wants it to be
visible to the system.

To set the schema only visible to the system, the user will have a
default setting. This is only work for OR. For the “AND” cases, we
always have other settings.

Move this out of visibleToConfig until we have real client need this.

Bug: 321326441
Test: VisibilityStoreTest
Relnote: Move mIsNotDisplayedBySystem to internalVisibilityConfig.
Change-Id: Ic1f39949de63cda0a5fb1453c8b912a3e34c7335