Add EdgeContentLayout2 as improved version of EdgeContentLayout in Material layouts
This layout has a couple of improvements:
* primary label should be at the fixed place on the top
  (that is the same as in PrimaryLayout) to give consistent look
  in the carousel.
  As opposed to EdgeContentLayout, where it's connected to the
  main content and vertically centered in the remaining space.
* vertical margins should be %, not fixed
* primary and secondary label should have side padding to ensure
  it doesn't go off the screen

Since there's many of these changes, we are deprecating the
existing EdgeContentLayout and creating a replacement version 2.

Fixes: 315950560
Test: Added
Relnote: "We have deprecated EdgeContentLayout and provided a new, improved replacement EdgeContentLayout2. This layout has a better placement of elements inside that helps with tiles consistency and better support for responsive content inset so that labels don't go off the screen edge."

Change-Id: Ia1cb2cee71591e315de75296b0587339fc10db21