Allow non full screen CircularProgressIndicator by removing margins
Currently, CircularProgressIndicator has 6dp margins all around it
so it's not tight to the screen. This is per UX spec, however,
this makes is difficult to be built in a smaller size, as it will
always have that extra space.

This adds a setter for removing those margins, to be used as a non
full screen variant. Note that, to control its size, it should be
added in a Box or other elements.

The EdgeContentLayout(2) also has margins around edges, but those
layouts are generally built as full screen layout and have many
other padding and margins based on % of the screen size, so
there's no need to update them in order to allow adding content
"inside" of the indicator, as that can be done manually.

Bug: 321972591
Test: Added
Relnote: "Addes method to remove outer margins from CircularProgressIndicator so it can be used as smaller component."
Change-Id: I55c060580dee3bf3719a340043faf81d91c930f2