Reduce code size at call sites of runtime checks
This change does two things. The first one is to replace calls to
stdlib check()/require() with "precondition" variants that assume
a String as the message lambda return type, and wrap the exception
in an extra function calls to avoid unnecessary inlining.

The second part of this change is a modification to runtimeCheck():
- Make the message lambda return String instead of Any, as this
  removes a toString() call at the call site we never need. This
  helps the instruction cache since we load fewer instructions.
  Also smaller binaries.
- runtimeCheck() wraps its exception in an extra function call,
  which is good. But this now causes the call sites to have
  extra instructions to throw a NothingValueException. Remove
  the Nothing return time from the wrapper function.

All together, the changes to runtimeCheck() yield a reduction of
7 DEX instructions and 5 aarch64 instructions *per call site*.

Test: compose-runtime tests suite
RelNote: N/A
Change-Id: Icdc4af3579a48a54af92c5cd686471a029b4848b