IdentityKey generation
Adds a class IdentityKey in its own package credentials.e2ee, with
a single static method .createFromPrf().
This API supports end-to-end messaging projects in the platform (e.g.
MLS). It is intended to
generate keys usable for signing from the PRF output of the Passkey PRF
Extension.

Relnote: Adds an API to generate IdentityKeys

Test: new unit tests + manual testing on P7

Change-Id: Iba31e191a7c1b74e7762613ca7abe8e69e37c83c