Fix final/open API breakages in Kruth
Note to API council - this is an unpublished library, so please ignore
this for API review. API tracking is only enabled to check against
compatibility with the public API of Truth.

Clean up the ABI breakages from Google Truth, that we "fixed" in Kruth
by making them final. Most of these APIs were likely intended to be
final, but it is more important to maintain compatibility in Kruth now.

Test: ./gradlew kruth:kruth:updateApi
Relnote: "N/A"
Change-Id: Id731a93562493873f5f1bb76d6431d36d1be4913