Add APIs to perform hit queries on Path
This change introduces two new APIs to query whether a given x/y
position lies inside a Path:
- Path.contains(Offset)
- PathHitTester(path).contains(Offset)

The former is easy and convenient but much more expensive if
multiple queries against the same path are required.
PathHitTester will cache the content of the path in a spatial
data structure to speed up subsequent queries.

This change does a few related things as well:
- Move Bézier functions from animation-core to graphics
- Move IntervalTree from animation-core to graphics
- Improve IntervalTree by halving the number of allocations
  required to build a tree of intervals
- Add the ability to load a Path from SVG and export it to SVG,
  which is convenient for testing and debugging
- Fix a bug in PathParser as it was inserting a superfluous
  "move to" command when closing a path

Test: PathHitTest, PathSvgTest, PathEasingTest
Relnote: Added PathHitTest and Path.contains(Offset) to check if a Path contains a specific point
Change-Id: I3b2186d2d3bacc037f56a1d6c2266083b78a6785