Fix issue with Switch devices and DateRangePicker
Removed a `clickable` modifier from the Month-Year text title to fix an
issue that was brought up. Seems like the use of clickable there is no
longer needed to get the additional actions when TalkBack is active and
we should not convay this element as actionable to the screen reader,
switch access and voice access users.

Bug: 323556934
Test: Manual
Relnote: "Fix an a11y issue with the DateRangePicker where the
Month-Year text title was conveyed as actionable to the screen reader,
Switch access, and Voice access users."

Change-Id: I2ac55010a9f9a39a8b0c3c76ed149488fba651d8