Improve performance for Activity startup via back dispatcher
Instead of using lifecycleScope for attaching the
onBackInvokedDispatcher we should go back to using a normal
LifecycleEventObserver as lifecycleScope comes with some additional
overhead from Kotlin coroutines.

We also now only call any of this code if you are actually using API 33
or above.

RelNote: "Fixed an Activity startup performance regression introduced in
the [Activity 1.9.0-alpha02] caused by calling the OnBackPressedDispatcher as part of the creation of the Activity."
Test: Tested with the Activity Startup benchmark
Bug: 322167634

Change-Id: Ie75e392c506ea3f142df175ae62dde7fd8220821