Make FocusDirectionWrapper a data class.
FocusDirectionWrapper instance name as being used as a test identifier and this messes up the test runtime dashboard for foundation test. In this change I made FocusDirectionWrapper a data class so it will use its parameter direction to implement the toString method and generate uniquely identifiable method names.

Test: Previous tests should pass.
Relnote: N/A
Change-Id: Ia2467de8849a20543bbf41097e180e822cea01f1