Remove unnecessary maxIntrinsicWidth call from TextFieldCoreModifier
This check was adopted from BasicTextFieldv1 while rewriting TextField scroll logic. It exists to make sure that intrinsically shorter texts do not extend to the size of the default text that's used by `Modifier.textFieldMinSize`. However, this is only true if `textFieldMinSize` modifier is applied to the right of scroll logic which happens to be in `Modifier.textFieldScroll` in `BasicTextField`. This is not true for `BasicTextField2`. We can safely delete this guard.

Test: BasicTextField2Test. Adopted from the original bug fix aosp/1706805.
Change-Id: I9d921835b53b58d48b35d47348aa3eb0809d1290