Semantics override for InputTransformation
With the addition of MaxTextLength semantics property, MaxLengthFilter should also be able to change the semantics of the `BasicTextField2` that it's attached to. Moreover, any custom `InputTransformation` may have different semantics implications. Rather than adding extra semantics modifiers to TextFields, `InputTransformation` should be able to influence the semantics organically.

Bug: 170648072
Test: gradle :compose:foundation:foundation:cAT
Relnote: "Add an optional `applySemantics` function to `InputTransformation` to influence the semantics of the `BasicTextField2` that it's applied to."
Change-Id: I74a2f4db178fe2660862749f10367bac57761ae9