Update supported feature API to be more extensible.
Add class to wrap the list of fold features. This is more extensible
since we will probably add properties and do not need a new class.
Move SupportedFoldFeature to be a top level class. Downcasting doesn't
seem to add any value since we need to check the vendor API level to use
the new class.

Relnote: Make SupportedWindowFeatures extensible.

Bug: 293340299
Test: ./gradlew window:extensions:extensions:test
Change-Id: I84754bc764f5a6b5b981df420cc65eded6b70803