Correct CameraGraph state transitions and collection
Currently, a regular CameraGraph might go through the following states:

GraphStateStopped --> GraphStateStarting --> GraphStateStarted -->
  GraphStateStopping --> GraphStateStopped

If users invoke CameraGraph.stop() before having a configured capture
session, we do not invoke GraphProcessor.onGraphStopped, and therefore
we don't transition to GraphStateStopped in that case.

Another issue is UseCaseCamera, when collecting states, cancels the
CoroutineScope from UseCaseThreads which actually belongs in a higher
@CameraScope, so it shouldn't be its job to cancel. Instead, we cancel
the CoroutineScope when CameraInternal is released.

Bug: 325088903
Test: None
Change-Id: I4ff1df1b43f307fce3a86176e656c0eb0b3a99b7