Add lint rule for PrimaryLayout responsiveness
When PrimaryLayout is used without setResponsiveContentInsetEnabled
or with passed false, the linter will report a warning at Builder ctor
location and provide a quick fix.

This is in order to promote usage of responsive PrimaryLayout that
fixes few issues with side space.

Fixes: 310856011
Test: Added
Relnote: "We have added a lint rule to report a warning when PrimaryLayout is used without setResponsiveContentInsetEnabled and provide a quick fix."

Change-Id: I120257c20dfa8912c2a0b4ecaabe843140d16285