Introduce GetScrollViewportLength semantic action.
In this CL we are adding the GetScrollViewportLength semantic action so we can pipe up information about the components being scrolled in compose to the a11y system. This CL also applies the usage of said property in Foundation Scrollable Lists.

Test: Add tests to check new behavior.
Relnote: In this CL we are adding the GetScrollViewportLength semantic action so we can pipe up information about the components being scrolled in compose to the a11y system. This CL also applies the usage of said property in Foundation Scrollable Lists.
Bug: 316558200
Change-Id: Ic5fa297df4613636529e12037a0b7d03bcacc534