Fix ExtraSupportedSurfaceCombinationsQuirk issue
Subset of guaranteed supported configurations are not guaranteed to be supported. aosp/2694107 was made to update the related logic. What we had verified on the added extra supported combination for Pixel series devices in ExtraSupportedSurfaceCombinationsQuirk is actually only (PRIV, PREVIEW) + (PRIV, ANALYSIS) + (YUV, MAXIMUM) which is a subset of Level-3 guaranteed supported configuration. The additional RAW stream is actually not used when we verified the streams combination. Therefore, we can remove the RAW stream from the quirk to make the Pixel series devices can use exactly-matched PRIV + PRIV + YUV combination without enabling stream sharing.

Bug: 324651953
Test: ./gradlew bOS
Change-Id: I1fc73772ac36b18b8db6ab27a8fed3a73b18295f