Use Access Key based authentication via Develocity Connector
This migrates the access to the Develocity Build Cache to use Access Key based authentication instead
of username and password authentication. This means that the access will use the same access key as is
is used for Build Scan publication.

Here is more information on the Develocity Connector [see here](https://docs.gradle.com/enterprise/gradle-plugin/#using_the_develocity_connector)