Implement NavArgument generator
Generates a list of NamedNavArgument from a serializer<T> where T is a serializable concerete class or object.

To convert arg types that are not natively supported by NavType, a map of arg KType to custom NavType is required. The custom NavTypes takes precedence over native NavTypes. For example, a custom NavType<Int> will be used for Int types instead of the official NavType.IntType.

Test: ./gradlew navigation:navigation-common:test
Bug: 188693139
Change-Id: I8043f09cead18c3874f1062427180e37be43f510