Add usercount to postview to ensure it is released when it is not in used
Postview DeferrableSurface is closed when unbinding use cases.
Without the usercount being incremented, the postview image reader
will be closed immediately when unbind is invoked.

We should increment usercount when it is in use, and decrement
when it is no longer needed to ensure it is released at the right time.

Bug: NA
Test: extensions test

Change-Id: Id5b5de8e9a2747e2328318817ad6370d07080f9d