Add SplitAttributes Builder from an existing instance
We sometimes need to update attributes in an existing SplitAttributes.
Before this change, we create a new Builder and manually set each
existing attributes. It is error-prone because when we add a new
attribute into SplitAttributes, we have to update every occurrence of
such uses or otherwise it will miss the new attribute in the created
SplitAttributes.

Bug: 293654166
Test: SplitAttributesTest
Relnote: "
  - Adding SplitAttributes Builder from an existing instance.
"

Change-Id: I0c8c7806898daf4bae85df19997a6db07e8275e0