Migrate Transaction binder and adapter to driver API.
To support coroutine transaction delegate in Java and Kotlin a new utility function was created in DBUtil that takes a suspend lambda, required when invoking the delegate super suspend function. Usage of Java 8 lambda was already present due to b/215557148, such codegen aspect is kept and not expanded to other writers.

Bug: 299168035
Test: BuilderTest + BundledSQLiteConnectionPoolTest
Change-Id: I60739bfa06d46571db5bf2ac5bdb1fd390ff8ee1