Extending Custom Entry API Surface w/ BiomPrompt
This adds to the CustomEntry the ability to utilize the single tap sign
in flow for providers.

This also cleans up our constructor build params to match the
'fromSlice' setup. Given how params can be out of order, this is a
cleaner approach and ensures we use the constructor that is meant to be
used (namely the internal one).

Bug: 325469890
Test: JUnit Tests chained
Relnote: "Allows enabling custom entries to use the single tap sign in
flow, where credential manager handles the biometirc prompt
authentication"

Change-Id: I060b1f3cdcf33b1087b7c83a0659f646be2f3dc6