Add FirExtensionRegistrar service for compose compiler plugin
Android Studio registers `ComposeIrGenerationExtension` to the project
EP area via a XML file (plugin.xml). However, it is not possible to do
the same for `ComposeFirExtensionRegistrar` because it does not have EP
area for project (we have to do that in code). In addition, we have to
run `configuration()` function for `FirExtensionRegistrar` in KT
compiler, so we have to access to `FirExtensionRegistrar` from IntelliJ.
(IntelliJ extension provider will deliver it to KT compiler)

To allow IntelliJ to access `ComposeFirExtensionRegistrar`, we want to
provider `ComposeFirExtensionRegistrar` using a `FirExtensionRegistrar`
service.

Note that we cannot use `ComposePluginRegistrar` for Android Studio
because `ComposePluginRegistrar` registers both
`ComposeFirExtensionRegistrar` and `ComposeIrGenerationExtension`,
which causes a duplicated registration of
`ComposeIrGenerationExtension`.

Bug: 325301896
Test: Manually tested compose compiler plugin with K2 LiveEdit on AS
Change-Id: I324b39cdad6a9fa101c1a972750fe31ad66d88a7