Add lint rule for EdgeContentLayout responsiveness
When EdgeContentLayout is used without setResponsiveContentInsetEnabled
or with passed false, the linter will report a warning at Builder ctor
location and provide a quick fix.

This is in order to promote usage of responsive and updated
EdgeContentLayout that fixes few issues with side space and primary
label positioning.

Bug: 324833434
Test: Added
Relnote: "We have added a lint rule to report a warning when EdgeContentLayout is used without setResponsiveContentInsetEnabled and provide a quick fix."
Change-Id: Ia1448b8180b5c95d177e3386ce7f5f7a288236b8