Correct nullability of vararg parameters that are not last
Previously, Metalava would treat Kotlin vararg parameters that were the
last parameter as non-null which is correct but would treat those that
were not last as nullable which is incorrect. Metalava has now been
fixed and this is updating the API signature files to reflect that
update.

Relnote: N/A
Bug: 325454211
Test: ./gradlew
Change-Id: Ib384d34ad73780863cae09d132a52850bfb7f764