Update baseline files to use correct element id for property
Previously, the element id used to identify the property item for which
an issue was reported was of a different structure to the element ids
used for other properties, like field. Metalava has been fixed and so
this change is simply updating the baseline files to match.

Updated using the following command:
    ./gradlew updateApiLintBaseline ignoreApiChanges

Bug: 327562507
Test: ./gradlew
Change-Id: I8e4c91ce0f8a6721ec711562c42158ce1ef57a02